% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_trigger_index_range.R
\name{calculate_trigger_index_range}
\alias{calculate_trigger_index_range}
\title{Calculate Trigger Index Range}
\usage{
calculate_trigger_index_range(x_range, sampling_frequency)
}
\arguments{
\item{x_range}{A integer vector that contains the upper and lower x-axis data limits that will c}

\item{sampling_frequency}{A double that represents the frequency the neuron data was sampled at.}
}
\value{
A integer vector that contains the trigger index ranges
}
\description{
A single spike will have mulitple triggers. Only a trigger point in needed. The other triggers within a certain range of the initial trigger can be 
discarded. This function calculates the range in which triggers will be discarded
}
\examples{
x_range <-c(-3,2)
sampling_frequency <- 10
trigger_index_range <- calculate_trigger_index_range(x_range,sampling_frequency)

}
