% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatSify.R
\docType{package}
\name{CatSify}
\alias{CatSify}
\alias{CatSify-package}
\alias{CatSify-package}
\title{CatSify: A package for Classification of Feline Neuronlogical Activity using Machine Learning}
\description{
The package provides 5 categories of important functions:

\itemize{
 \item{"extract_data"}
 \item{"process_data"}
 \item{"generate_features"}
 \item{"conduct_machine_learing"}
 \item{"output"}
}
}
\section{CatSify functions}{

The foo functions ...
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jdearmas/intro_to_R}
}

}
\author{
\strong{Maintainer}: John De Armas \email{dearmas.john1996@gmail.com}

}
