% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_std_plot.R
\name{create_std_plot}
\alias{create_std_plot}
\title{Create Plot of Standard Deviations of Data}
\usage{
create_std_plot(xdata_range, ydata, desired_std_range, range_of_std_of_data)
}
\arguments{
\item{xdata_range}{An integer vector of the X-axis values}

\item{ydata}{An integer vector of the Y-axis values}

\item{desired_std_range}{An integer vector of the range of standard deviations that wish to be displayed}

\item{range_of_std_of_data}{An integer vector that contains the values of each standard deviation range}
}
\value{
A line plot with horizontal standard deviation lines
}
\description{
Create a plot with horizontal lines that represent the standard deviations
}
\examples{
xdata_range <- 1:100
ydata <- sin(xdata_range)
desired_std_range <- -1:1
range_of_std_of_data <- c(-0.5,0,0.5)
create_std_plot(xdata_range,ydata,desired_std_range,range_of_std_of_data)
![](create_std_plot_example.png "Example Plot Title")
if{html}{figure{create_std_plot_example.png}{options: width="60\%" alt="Figure: workspace_id.png"}}
}
