% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_initial_trigger_indices.R
\name{extract_initial_trigger_indices}
\alias{extract_initial_trigger_indices}
\title{Extract Initial Trigger Indices}
\usage{
extract_initial_trigger_indices(trigger_index_range, trigger_indices)
}
\arguments{
\item{trigger_index_range}{A integer vector; Represents the range of trigger indices 
from the first that will be grouped into one response}

\item{trigger_indices}{An integer vector; All the indices of the neuron data that met or surpassed the trigger condition.}
}
\value{
An integer vector; An array of the indices of first value of each response
}
\description{
The trigger indices obtained by \strong{extract_trigger_indices} 
contains every value on the spike that met or surpassed the trigger condition.
Only the first trigger index of each response matters.
}
\examples{
trigger_index_range <- c(-2,2)
trigger_indices <- c(10,11,13,14,16)
initial_trigger_indices  <- extract_initial_trigger_indices(trigger_index_range,trigger_indices)
}
