% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_responses.R
\name{extract_responses}
\alias{extract_responses}
\title{Extract Responses}
\usage{
extract_responses(xdata, ydata, x_range, trigger_indices)
}
\arguments{
\item{xdata}{An integer vector; The time values of the}

\item{ydata}{An integer vector; The values of the neuron data}

\item{x_range}{An integer vector; The range of time from the trigger point that defines the start and end of the response}

\item{trigger_indices}{An integer vector; The points in time that the neuron data hits the trigger conditions}
}
\value{
An integer vector; A neuron response
}
\description{
The neuron responses in the binary data are extracted and recorded
}
\examples{
bindata <- import_binary_data_file('Data/Cat_Data.bin',total_file_length = 10)
time <- time(bindata)
x_range <- c(-1,1)
trigger_indices <- 5
neuron_response <- extract_responses(time,bindata,x_range,trigger_indices)
}
