% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trigger_indices.R
\name{extract_trigger_indices}
\alias{extract_trigger_indices}
\title{Extract Trigger Indices}
\usage{
extract_trigger_indices(data, trigger_condition)
}
\arguments{
\item{data}{A integer vector of data values}

\item{trigger_condition}{An integer vector that represent the condition that must be meet in order for the trigger indices to be recorded.}
}
\value{
trigger_indices An integer vector of indices of the \strong{data} that met or surpassed the \strong{trigger condition}
}
\description{
The data is reviewed and if it activites the trigger condition, the index of the data that caused the trigger is recorded.
}
\examples{
data <- sin(seq(1,pi,100));
trigger_condition <- c(1,3);
trigger_indices <- trigger_condition_indices(data,trigger_condition) 
}
