% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_binary_data_file.R
\name{import_binary_data_file}
\alias{import_binary_data_file}
\title{Import a binary data file}
\usage{
import_binary_data_file(bin_file_path, total_file_length)
}
\arguments{
\item{bin_file_path}{A character value that contains the full path to the binary data file one wishes to import}

\item{total_file_lenght}{A double that represents the number of values you from the binray data file you wish to import}
}
\description{
More Details
}
\examples{
bin_file_path <- "C:/Users/John/Data.bin"
total_file_length <- 10
bindata <- import_binary_data_file(bin_file_path,total_file_length) 
}
