% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_bin_data.R
\name{preview_bin_data}
\alias{preview_bin_data}
\title{Preview Binary Data}
\usage{
preview_bin_data(bindata, preview_range)
}
\arguments{
\item{bindata}{An integer vector of the neuron data}

\item{preview_range}{An integer vector that represents the x-axis}
}
\value{
A plot of the inputed data over the preview range
}
\description{
This function outputs a plot of the binary data
}
\examples{
bindata <- import_binary_data_file('Data/Cat_Data.bin',10)
preview_range <- 1:10
preview_bin_data(bindata,preview_range)
}
